+++ src/common/Channel.cpp
@@ -28,10 +28,9 @@ Channel::Channel(const QString &name, Type type)
     : completionModel(*this)
     , lastDate_(QDate::currentDate())
     , name_(name)
+    , messages_(getSettings()->twitchMessageLogLimit)
     , type_(type)
 {
-    this->messages_ = std::make_shared<LimitedQueue<MessagePtr>>(
-        getSettings()->twitchMessageLogLimit);
 }
 
 Channel::~Channel()
@@ -71,12 +70,12 @@ bool Channel::isEmpty() const
 
 bool Channel::hasMessages() const
 {
-    return !this->messages_->empty();
+    return !this->messages_.empty();
 }
 
 LimitedQueueSnapshot<MessagePtr> Channel::getMessageSnapshot()
 {
-    return this->messages_->getSnapshot();
+    return this->messages_.getSnapshot();
 }
 
 void Channel::addMessage(MessagePtr message,
@@ -101,7 +100,7 @@ void Channel::addMessage(MessagePtr message,
         app->logging->addMessage(this->name_, message, channelPlatform);
     }
 
-    if (this->messages_->pushBack(message, deleted))
+    if (this->messages_.pushBack(message, deleted))
     {
         this->messageRemovedFromStart.invoke(deleted);
     }
@@ -225,7 +224,7 @@ void Channel::disableAllMessages()
 void Channel::addMessagesAtStart(std::vector<MessagePtr> &_messages)
 {
     std::vector<MessagePtr> addedMessages =
-        this->messages_->pushFront(_messages);
+        this->messages_.pushFront(_messages);
 
     if (addedMessages.size() != 0)
     {
@@ -235,7 +234,7 @@ void Channel::addMessagesAtStart(std::vector<MessagePtr> &_messages)
 
 void Channel::replaceMessage(MessagePtr message, MessagePtr replacement)
 {
-    int index = this->messages_->replaceItem(message, replacement);
+    int index = this->messages_.replaceItem(message, replacement);
 
     if (index >= 0)
     {
@@ -245,7 +244,7 @@ void Channel::replaceMessage(MessagePtr message, MessagePtr replacement)
 
 void Channel::replaceMessage(size_t index, MessagePtr replacement)
 {
-    if (this->messages_->replaceItem(index, replacement))
+    if (this->messages_.replaceItem(index, replacement))
     {
         this->messageReplaced.invoke(index, replacement);
     }
@@ -264,7 +263,7 @@ MessagePtr Channel::findMessage(QString messageID)
 {
     MessagePtr res;
 
-    if (auto msg = this->messages_->rfind([&messageID](const MessagePtr &msg) {
+    if (auto msg = this->messages_.rfind([&messageID](const MessagePtr &msg) {
             return msg->id == messageID;
         });
         msg)